ssh-agent called only if there is an id_rsa file

+++ .bashrc
@@ -4,13 +4,15 @@
 
 # ssh-agent {{{
 
-if [ ! -S ~/.ssh/ssh_auth_sock ]; then
-    eval $(ssh-agent)
-    ln -sf "$SSH_AUTH_SOCK" ~/.ssh/ssh_auth_sock
-fi
+if [ -f ~/.ssh/id_rsa ]; then
+    if [ ! -S ~/.ssh/ssh_auth_sock ]; then
+        eval $(ssh-agent)
+        ln -sf "$SSH_AUTH_SOCK" ~/.ssh/ssh_auth_sock
+    fi
 
-export SSH_AUTH_SOCK=~/.ssh/ssh_auth_sock
-ssh-add -l > /dev/null || ssh-add
+    export SSH_AUTH_SOCK=~/.ssh/ssh_auth_sock
+    ssh-add -l > /dev/null || ssh-add
+fi
 
 # }}}
 # aliases {{{
